Changes Animale.java AnimaleDto.java AnimaleRepository.java

+++ animali/animali/src/main/java/it/adozioni/animali/Repository/AnimaleRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AnimaleRepository extends JpaRepository<AnimaleRepository,Long> {
+public interface AnimaleRepository extends JpaRepository<AnimaleRepository,Integer> {
     Animale findByNome(String nome);
     List<Animale>findByRazza(String razza);
     List<Animale>findByGenere(String genere);
